ubah crud daftar menu

+++ src/com/unindra/restoserver/controllers/DaftarMenuController.java
@@ -3,6 +3,7 @@ package com.unindra.restoserver.controllers;
 import com.jfoenix.controls.*;
 import com.jfoenix.controls.datamodels.treetable.RecursiveTreeObject;
 import com.unindra.restoserver.Dialog;
+import com.unindra.restoserver.models.DetailRamen;
 import com.unindra.restoserver.models.Level;
 import com.unindra.restoserver.models.Menu;
 import javafx.collections.FXCollections;
@@ -13,9 +14,14 @@ import javafx.scene.control.TreeItem;
 import javafx.scene.control.TreeTableColumn;
 import javafx.scene.control.TreeTableView;
 import javafx.scene.input.MouseEvent;
+import javafx.scene.layout.HBox;
+import javafx.stage.FileChooser;
 import javafx.stage.Stage;
 
+import java.io.File;
+import java.io.IOException;
 import java.net.URL;
+import java.nio.file.Files;
 import java.util.ResourceBundle;
 
 import static com.unindra.restoserver.models.Level.getLevels;
@@ -33,10 +39,13 @@ public class DaftarMenuController implements Initializable {
     public JFXTreeTableView<Level> levelTableView;
     public JFXTextField hargaLevelField;
     public JFXTextField levelField;
+    public HBox formForRamenPane;
+    public JFXButton pilihGambarButton;
 
     private ObservableList<String> tipeList;
     private Menu menu;
     private Level level;
+    private DetailRamen detailRamen;
 
     @Override
     public void initialize(URL location, ResourceBundle resources) {
@@ -55,7 +64,7 @@ public class DaftarMenuController implements Initializable {
         menuTableView.getColumns().add(hargaCol);
         menuTableView.setColumnResizePolicy(TreeTableView.CONSTRAINED_RESIZE_POLICY);
 
-        tipeList = FXCollections.observableArrayList("minuman", "cemilan", "lainnya");
+        tipeList = FXCollections.observableArrayList("ramen","minuman", "cemilan", "lainnya");
         tipeComboBox.setItems(tipeList);
 
         TreeTableColumn<Level, Integer> levelCol = new TreeTableColumn<>("Level");
@@ -85,23 +94,41 @@ public class DaftarMenuController implements Initializable {
 
     public void actionHandle() {
         if (actionButton.getText().equals("Tambah")) {
-            Menu menu = new Menu(
+            menu = new Menu(
                     namaField.getText(),
                     tipeComboBox.getSelectionModel().getSelectedItem(),
                     Integer.valueOf(hargaField.getText()));
-            if (menu.add()) {
-                getDialog().information("Berhasil!", "Menu berhasil ditambahkan");
-                reset();
+            if (menu.getTipe().equals("ramen")) {
+                if (detailRamen != null) {
+                    detailRamen.setNama_menu(namaField.getText());
+                    detailRamen.setDeskripsi(deskArea.getText());
+                    if (menu.add() && detailRamen.add()) {
+                        getDialog().information("Berhasil!", "Menu berhasil ditambahkan");
+                        reset();
+                    } else getDialog().information("Gagal", "Menu gagal ditambahkan");
+                } else getDialog().information("Gagal", "Gambar belum dipilih");
+            } else {
+                if (menu.add()) {
+                    getDialog().information("Berhasil!", "Menu berhasil ditambahkan");
+                    reset();
+                } else getDialog().information("Gagal", "Menu gagal ditambahkan");
             }
-            else getDialog().information("Gagal", "Menu gagal ditambahkan");
-        } else {
+        } else { // Ubah
             menu.setNama_menu(namaField.getText());
             menu.setTipe(tipeComboBox.getSelectionModel().getSelectedItem());
             menu.setHarga_menu(Integer.valueOf(hargaField.getText()));
-            if (menu.update()) {
-                getDialog().information("Berhasil!", "Menu berhasil diubah");
-                reset();
-            } else getDialog().information("Gagal", "Menu gagal diubah");
+            if (menu.getTipe().equals("ramen")) {
+                detailRamen.setDeskripsi(deskArea.getText());
+                if (menu.update() && detailRamen.update()) {
+                    getDialog().information("Berhasil!", "Menu berhasil diubah");
+                    reset();
+                } else getDialog().information("Gagal", "Menu gagal diubah");
+            } else {
+                if (menu.update()) {
+                    getDialog().information("Berhasil!", "Menu berhasil diubah");
+                    reset();
+                } else getDialog().information("Gagal", "Menu gagal diubah");
+            }
         }
     }
 
@@ -116,10 +143,14 @@ public class DaftarMenuController implements Initializable {
     public void pilihHandle(MouseEvent mouseEvent) {
         if (!menuTableView.getSelectionModel().isEmpty()) {
             menu = menuTableView.getSelectionModel().getSelectedItem().getValue();
+            detailRamen = DetailRamen.detailRamen(menu);
+
             namaField.setText(menu.getNama_menu());
             int index = tipeList.indexOf(menu.getTipe());
             tipeComboBox.getSelectionModel().clearAndSelect(index);
             hargaField.setText(String.valueOf(menu.getHarga_menu()));
+            if (detailRamen != null) deskArea.setText(detailRamen.getDeskripsi());
+
             titleLabel.setText("UBAH MENU");
             namaField.setDisable(true);
             hapusButton.setVisible(true);
@@ -145,10 +176,17 @@ public class DaftarMenuController implements Initializable {
                 "Anda yakin ingin menghapus menu ini?",
                 event -> {
                     Menu menu = menuTableView.getSelectionModel().getSelectedItem().getValue();
-                    if (menu.delete()) {
-                        alert.information("Berhasil!", "Menu berhasil dihapus");
-                        reset();
-                    } else alert.information("Gagal", "Menu gagal dihapus");
+                    if (menu.getTipe().equals("ramen")) {
+                        if (menu.delete() && detailRamen.delete()) {
+                            alert.information("Berhasil!", "Menu berhasil dihapus");
+                            reset();
+                        } else alert.information("Gagal", "Menu gagal dihapus");
+                    } else {
+                        if (menu.delete()) {
+                            alert.information("Berhasil!", "Menu berhasil dihapus");
+                            reset();
+                        } else alert.information("Gagal", "Menu gagal dihapus");
+                    }
                 });
     }
 
@@ -168,12 +206,36 @@ public class DaftarMenuController implements Initializable {
         hargaLevelField.setDisable(true);
         levelTableView.getSelectionModel().clearSelection();
         namaField.requestFocus();
+        pilihGambarButton.setText("Pilih gambar... (max : 2048 KB)");
+    }
+
+    public void tipeHandle() {
+        if (tipeComboBox.getSelectionModel().getSelectedItem() != null) {
+            if (tipeComboBox.getSelectionModel().getSelectedItem().equals("ramen")) formForRamenPane.setDisable(false);
+            else {
+                formForRamenPane.setDisable(true);
+                detailRamen = null;
+                deskArea.setText("");
+                pilihGambarButton.setText("Pilih gambar... (max : 2048 KB)");
+            }
+        }
+    }
+
+    public void pilihGambarHandle() throws IOException {
+        FileChooser fileChooser = new FileChooser();
+        fileChooser.setTitle("Open Resource File");
+        File file = fileChooser.showOpenDialog(actionButton.getScene().getWindow());
+        if (file != null) {
+            if (file.length() <= 2048000) {
+                pilihGambarButton.setText(file.getName());
+                detailRamen = new DetailRamen(Files.readAllBytes(file.toPath()));
+            } else {
+                getDialog().information("Gagal", "Ukuran foto terlalu besar");
+            }
+        }
     }
 
     private Dialog getDialog() {
         return new Dialog((Stage) actionButton.getScene().getWindow());
     }
-
-
-
 }
